refactor: homepage search tests to dedicated test files
